% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_ModifiedConditional.R
\name{CI_ModifiedConditional}
\alias{CI_ModifiedConditional}
\title{CI_ModifiedConditional Function}
\usage{
CI_ModifiedConditional(
  deltahat,
  deltaSigma,
  Al,
  Au,
  alpha = 0.05,
  alphac = alpha * 0.8,
  eta = 10^-3,
  B = 4000,
  Blarge = B * 10,
  tol = 10^-3,
  tol_r = 10^-3
)
}
\arguments{
\item{deltahat}{A  p*1 vector, see Assumption 1.}

\item{deltaSigma}{A p*p matrix, see Assumption 1.}

\item{Al}{A B*p matrix, see Assumption 1.}

\item{Au}{A B*p matrix, see Assumption 1.}

\item{alpha}{The nominal rejection rate.}

\item{alphac}{A tuning parameter, suggested value is alphac = 0.8*alpha.}

\item{eta}{A tuning parameter, suggested value is 0.001, see Bei(2024) eq(40).}

\item{B}{The number of bootstrap draws used to caluclate c^t, suggested value 4000.}

\item{Blarge}{The number of bootstrap draws used to caluclate c^t, suggested value B*10.}

\item{tol}{Tolerance level, suggested value 10^-3.}

\item{tol_r}{Tolerance level, suggested value 10^-3.}
}
\value{
A list with ConfidenceInterval, delta_opt, c_t.
}
\description{
This function calculates the CI base on Bei(2024) "Inference on Union Bounds"
}
