% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_RM.R
\name{CI_RM}
\alias{CI_RM}
\title{CI_RM Function}
\usage{
CI_RM(
  betahat,
  betaSigma,
  prePeriodIndices,
  ell_post,
  M = 1,
  alpha = 0.05,
  alphac = 0.8 * alpha,
  eta = 0.001,
  B = 4000,
  Blarge = 40000,
  tol = 0.001,
  tol_r = 0.001
)
}
\arguments{
\item{betahat}{T by 1, regression coefficients, e.g. in Rambachan and Roth(2023) Equation 2.}

\item{betaSigma}{T by T, covariance matrix of betahat.}

\item{prePeriodIndices}{T_pre by 1, e.g. (1,2,3,4)}

\item{ell_post}{T_post by 1, t(ell)*tau_post is the parameter of interest}

\item{M}{1 by 1, level of relaxation.}

\item{alpha}{1 by 1, 1-alpha is the confidence level}

\item{alphac}{1 by 1, a tuning parameter.}

\item{eta}{1 by 1, a tuning parameter, see Bei(2024) Equation 40.}

\item{B}{The number of bootstrap draws used to caluclate c^t}

\item{Blarge}{The number of bootstrap draws used to caluclate c^t.}

\item{tol}{Tolerance level}

\item{tol_r}{Tolerance level, suggested value 10^-3.}
}
\value{
A list with ConfidenceInterval, delta_opt, c_t.
}
\description{
This function calculates confidence intervals for the treatment effect under relative magnitudes relaxation.
}
