% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_SDRM.R
\name{CI_SDRM}
\alias{CI_SDRM}
\title{CI_SDRM Function}
\usage{
CI_SDRM(
  betahat,
  betaSigma,
  prePeriodIndices,
  ell_post,
  M = 1,
  alpha = 0.05,
  alphac = 0.8 * alpha,
  eta = 0.001,
  B = 4000,
  Blarge = 40000,
  tol = 0.001,
  tol_r = 0.001
)
}
\arguments{
\item{betahat}{T by 1, difference-in-differences coefficients, e.g. in Rambachan and Roth(2023, REStud) Equation 2.}

\item{betaSigma}{T by T, covariance matrix of betahat.}

\item{prePeriodIndices}{T_pre by 1, e.g. (1,2,3,4)}

\item{ell_post}{T_post by 1, ell\%*\%tau_post is the parameter of interest, see Rambachan and Roth(2023, REStud) Section 2.3.}

\item{M}{1 by 1, level of relaxation, see Rambachan and Roth (2023, REStud) Section 2.4.1.}

\item{alpha}{1 by 1, 1-alpha is the confidence level}

\item{alphac}{1 by 1, a tuning parameter, see Bei(2024) Section 3.2.}

\item{eta}{1 by 1, a tuning parameter, see Bei(2024) Equation 40.}

\item{B}{The number of bootstrap draws used to caluclate c^t}

\item{Blarge}{The number of bootstrap draws used to caluclate c^t.}

\item{tol}{Tolerance level to calculate the confidence interval.}

\item{tol_r}{Tolerance level to check whether |rho| = 1}
}
\value{
A list with ConfidenceInterval, c_MC_l, c_MC_u, c_t, delta_opt
}
\description{
Constructs robust confidence intervals for the difference-in-differences estimand under second differences relative magnitudes relaxation, see Rambachan and Roth (2023, REStud) Section 2.4.5.
}
\examples{
# Load example data
# the sample data is from Dustmann et al (2022), see Bei(2024) Section 5
data(example_data, package = "UnionBounds")
betahat <- example_data$betahat
betaSigma <- example_data$betaSigma
prePeriodIndices <- example_data$prePeriodIndices

# Call the CI_SDRM function with the example data
ell_post <- c(1, 0)  # the parameter of interest is the treatment effect at time 1
CI <- CI_SDRM(betahat, betaSigma, prePeriodIndices, ell_post, M = 1, alpha = 0.05)
print(CI$ConfidenceInterval) # This is the SDRM confidence interval

}
