% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIcon.R
\name{CIcon}
\alias{CIcon}
\title{CIcon Function}
\usage{
CIcon(deltahat, deltaSigma, Al, Au, alphac, CI_p, tol, tol_r)
}
\arguments{
\item{deltahat}{A numeric vector.}

\item{deltaSigma}{A matrix.}

\item{Al}{A matrix.}

\item{Au}{A matrix.}

\item{alphac}{A numeric value.}

\item{CI_p}{A numeric vector.}

\item{tol}{A numeric value.}

\item{tol_r}{A numeric value.}
}
\value{
A numeric vector representing the confidence interval.
}
\description{
This function calculates the conditional confidence interval.
}
