% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIcon_TNbounds.R
\name{CIcon_TNbounds}
\alias{CIcon_TNbounds}
\title{CIcon_TNbounds Function}
\usage{
CIcon_TNbounds(
  theta,
  deltahat,
  Al,
  Au,
  sigma_l,
  sigma_u,
  corr_m,
  corr_l,
  corr_u,
  tol_r
)
}
\arguments{
\item{theta}{A numeric value. H0.}

\item{deltahat}{A B*p matrix.}

\item{Al}{A matrix.}

\item{Au}{A matrix.}

\item{sigma_l}{A numeric vector.}

\item{sigma_u}{A numeric vector.}

\item{corr_m}{A matrix.}

\item{corr_l}{A matrix.}

\item{corr_u}{A matrix.}

\item{tol_r}{A numeric value.}
}
\value{
A list with th_1 and th_2.
}
\description{
This function calculates the bounds of the truncated Normal.
}
