% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIproj_p.R
\name{CIproj_p}
\alias{CIproj_p}
\title{CIproj_p Function}
\usage{
CIproj_p(
  c,
  delta,
  alphac,
  deltastar_demean,
  Al,
  Au,
  sigma_l,
  sigma_u,
  corr_m,
  corr_l,
  corr_u,
  eta,
  tol_r
)
}
\arguments{
\item{c}{A numeric value. The lower truncation.}

\item{delta}{A numeric vector. The true value of delta.}

\item{alphac}{A numeric value.}

\item{deltastar_demean}{A matrix.}

\item{Al}{A B*p matrix, see Assumption 1.}

\item{Au}{A B*p matrix, see Assumption 1.}

\item{sigma_l}{A numeric vector.}

\item{sigma_u}{A numeric vector.}

\item{corr_m}{A matrix defined in Bei(2024) Lemma 1.}

\item{corr_l}{A matrix defined in Bei(2024) Lemma 1.}

\item{corr_u}{A matrix defined in Bei(2024) Lemma 1.}

\item{eta}{A numeric value.}

\item{tol_r}{A numeric value.}
}
\value{
A numeric value.
}
\description{
This function calculates the rejection rate given true value delta and lower truncation c
}
