% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example Data}
\format{
A list containing:
\describe{
\item{example_data}{betahat, betaSigma and prePeriodIndices}
}
}
\usage{
data(example_data)
}
\description{
This dataset contains example data for betahat, betaSigma, and prePeriodIndices.
}
\keyword{datasets}
