% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_ModifiedConditional.R
\name{CI_ModifiedConditional}
\alias{CI_ModifiedConditional}
\title{CI_ModifiedConditional Function}
\usage{
CI_ModifiedConditional(
  deltahat,
  deltaSigma,
  Al,
  Au,
  alpha = 0.05,
  alphac = alpha * 0.8,
  eta = 10^-3,
  B = 4000,
  Blarge = B * 10,
  tol = 10^-3,
  tol_r = 10^-3
)
}
\arguments{
\item{deltahat}{p by 1 vector, see Bei(2024) Assumption 1.}

\item{deltaSigma}{p by p matrix, see Bei(2024) Assumption 1.}

\item{Al}{\eqn{|\mathcal{B}|} by p matrix, see Bei(2024) Assumption 1.}

\item{Au}{\eqn{|\mathcal{B}|} by p matrix, see Bei(2024) Assumption 1.}

\item{alpha}{The nominal rejection rate.}

\item{alphac}{A tuning parameter, suggested value is alphac = 0.8*alpha, see Bei(2024) Section 3.2.}

\item{eta}{A tuning parameter, suggested value is 0.001, see Bei(2024) eq(40).}

\item{B}{The number of bootstrap draws used to caluclate c^t, suggested value 4000.}

\item{Blarge}{The number of bootstrap draws used to caluclate c^t, suggested value B*10.}

\item{tol}{Tolerance level to do the grid search, suggested value 10^-3.}

\item{tol_r}{Tolerance level to determine |rho|=1, suggested value 10^-3.}
}
\value{
A list with ConfidenceInterval, c_MC_l, c_MC_u, c_t, delta_opt.
}
\description{
Calculates the confidence interval for \eqn{\theta \in \left[\min_{b \in \mathcal{B}} \lambda_{\ell,b}, \max_{b \in \mathcal{B}} \lambda_{u,b}\right]} using the modified conditional inference proposed by Bei(2024) "Inference on Union Bounds".
Assume that the estimators \eqn{\hat{\lambda_{\ell}} = A_{\ell}\hat{\delta}} and \eqn{\hat{\lambda_{u}} = A_{u}\hat{\delta}}. \eqn{\hat{\delta}} is asymptotically normal with covariance matrix \eqn{deltaSigma}.
}
\examples{
# Example:  min{delta_1, delta_2} <= theta <= max{delta_3, delta_2}
set.seed(0)
deltaSigma <- matrix(c(1, 0, 0, 0, 1, 0, 0, 0, 1), nrow = 3, ncol = 3, byrow = TRUE)
deltahat <- MASS::mvrnorm(n = 1, c(0,0, 0), deltaSigma)
deltahat <- matrix(deltahat, nrow = 3, ncol = 1)
Al <- matrix(c(1, 0, 0, 0, 1, 0), nrow = 2, ncol = 3, byrow = TRUE)
Au <- matrix(c(0, 0, 1, 0, 1, 0), nrow = 2, ncol = 3, byrow = TRUE)
# Call the CI_ModifiedConditional function
CI <- CI_ModifiedConditional(deltahat, deltaSigma, Al, Au, alpha = 0.05)
print(CI$ConfidenceInterval) # This is the confidence interval
print(CI$c_MC_l) # The lower bound of CI is min{lambda_{l,b} - c_MC_l*sigma_l}
print(CI$c_MC_u) # The upper bound of CI is min{lambda_{u,b} - c_MC_u*sigma_l}
print(CI$c_t) # c_t is the lower bound of c_MC_l and c_MC_u, see Bei(2024) Section 3.3
print(CI$delta_opt) # the maximizer in Bei(2024) eq(26)
}
